# Backend/MivoraSwagger.yaml
openapi: 3.0.4
info:
  title: Mivora - 2025
  description: |-
    This OpenAPI spec describes the backend API for Mivora, a platform that handles users, events, tickets, and real-time event messaging. It includes full user account flows such as registration, login, logout, token refresh, profile updates, avatar uploads, email verification, and password-reset procedures. Event organizers can create, update, publish, cancel, and manage event posters, while both organizers and attendees can send and fetch event messages. Users can browse published events, view details, and book tickets. The ticketing system supports booking, listing with filters, retrieving individual tickets, canceling or checking in tickets, and QR-code-based scanning. All sensitive and user-specific operations use JWT bearer authentication, and list endpoints follow a consistent pagination pattern.
    some useful links:
    - [The Mivora repository](https://github.com/SantiagoDeStefano/Mivora)
    - [The source API definition for Mivora](https://github.com/SantiagoDeStefano/Mivora/blob/main/Backend/MivoraSwagger.yaml)

  version: 0.0.1
servers:
  - url: /
    description: Same host as Swagger
  - url: http://47.129.118.52:4000
    description: EC2 server
  - url: http://localhost:4000
    description: Local server
  - url: http://26.35.82.76:4000
    description: LAN server
  - url: https://khoinguyenpham.name.vn
    description: Production api server
...
paths:
...
  /api/v1/users/refresh-token:
    post:
      tags:
        - users
      summary: Refreshes the access token
      operationId: refreshToken
      requestBody:
        description: The refresh token to be used for generating a new access token
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjg0NmE5MjJjOGQ2YzYwYzI5ZWM5MmQzIiwidG9rZW5fdHlwZSI6MSwidmVyaWZ5IjoxLCJpYXQiOjE3NTI1ODk5MzYsImV4cCI6MTc2MTIyOTkzNn0.C4W_Me1Di-owhrpj1VJ6j-ImoPtWMy6QMPT6R-uJ4a8'
      responses:
        '200':
          description: Get new access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: New tokens returned successfully
                  result:
                    $ref: '#/components/schemas/AuthenticateSuccessfully'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Used refresh token or does not exist'
...
  /api/v1/tickets:
    post:
      tags:
        - ticket
      summary: Book a ticket to an event
      operationId: bookTicket
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event_id:
                  type: string
                  format: uuid
                  example: ba6e6578-da2d-4317-a605-279f9c207228
      responses:
        '200':
          description: Book ticket successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Book ticket successfully
                  result:
                    type: object
                    properties:
                      ticket:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
                            example: c341339e-6277-464a-a35c-7a09184ca600
                          event_title:
                            type: string
                            example: Grass-roots dedicated hierarchy
                          event_status:
                            type: string
                            enum: [booked, canceled, checked_in]
                            example: published
                          ticket_status:
                            type: string
                            enum: [booked, canceled, checked_in]
                            example: booked
                          checked_in_at:
                            type: string
                            format: date-time
                            nullable: true
                            example: null
                          price_cents:
                            type: integer
                            example: 31315
                          qr_code:
                            type: string
                            description: Ticket QR code as data URL (base64 image)
                            example: data:image/png;base64,<base64-data>
        '403':
          description: Event's creator cannot book ticket
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Event's creator cannot book ticket
        '409':
          description: One user per event only
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: One user per event only
...
  /api/v1/tickets/check-ins:
    post:
      tags:
        - ticket
      summary: Scan (check-in) a ticket by qr_code_token
      operationId: scanTicket
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                qr_code_token:
                  type: string
                  format: uuid
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJldmVudF9pZCI6ImE4ZDI5ZWNkLTAzYWUtNGI2MS1hOThiLTMxYmNjMmY0NGMxNyIsInVzZXJfaWQiOiIwNzk2NWI3Mi0xZDU3LTRkM2EtOWZjYy01MmRjNWRjZGFhMzkiLCJ0b2tlblR5cGUiOjQsImlhdCI6MTc2NDgzMTkyMX0.vPLtSgO0tlyFAzQ_wmsSUudb5M7d9pznYtwKxjYB60I
      responses:
        '200':
          description: Ticket scanned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Ticket scanned successfully
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Ticket not found
        '403':
          description: Ticket has been checked in already
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Ticket has been checked in already
        '409':
          description: Ticket already checked in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Ticket already checked in
components:
  schemas:
...
    AuthenticateSuccessfully:
      type: object
      properties:
        result:
          type: object
          properties:
            access_token:
              type: string
              example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjg0NmE5MjJjOGQ2YzYwYzI5ZWM5MmQzIiwidG9rZW5fdHlwZSI6MCwidmVyaWZ5IjoxLCJpYXQiOjE3NTI1ODk5MzYsImV4cCI6MTc1MjU4OTk2Nn0.N0NpQZDcu-gMpXLApVimd0etDPJoz7cp6ksGsdbP1WA'
            refresh_token:
              type: string
              example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjg0NmE5MjJjOGQ2YzYwYzI5ZWM5MmQzIiwidG9rZW5fdHlwZSI6MSwidmVyaWZ5IjoxLCJpYXQiOjE3NTI1ODk5MzYsImV4cCI6MTc2MTIyOTkzNn0.C4W_Me1Di-owhrpj1VJ6j-ImoPtWMy6QMPT6R-uJ4a8'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Use the `Bearer` scheme to authenticate requests. Include the token in the `Authorization` header as follows:
        ```
        Authorization: Bearer <your_token>
        ```